MINOR fu changes
Ryan identified a mismatch between our understanding of the messaging
feature, while we will definitely need another meeting to make sure we
are on the same page, I looked at the fu side and discovered the
sendmsghandler was not adhering to the sendMessage interface.  I matched
fu to the interface and added some comments (bad habit of mine not to
comment).